+++ R/get_files.R
@@ -10,12 +10,11 @@
 #'   all R files will be analysed
 #' @param pattern A character string containing a regular expression to be
 #'   matched in all R files found in the specified directory
+#' @param suffix A string with a pattern to filter the searched file endings.
+#'   The default is ".R$" to only look for R-files. Use "." to search all files.
 #'
 #' @return Returns a character vector containing all files that match the
 #'   pattern
-#'
-#' @importFrom magrittr %>%
-#' @importFrom readr read_file
 #'
 #' @export
 #'
@@ -26,19 +25,30 @@
 #' # Find all files in . that load packages through library
 #' get_files(dir = ".", pattern = "library")
 #' }
-get_files <- function(dir, pattern = "") {
+get_files <- function(dir, pattern = "", suffix = ".R$") {
 
   if (!is.character(dir)) stop("'dir' must be a character string")
   if (!is.character(pattern)) stop("'pattern' must be a character string")
   if (!dir.exists(dir)) stop(sprintf("Directory '%s' doesn't exist.", dir))
 
-  files <- list.files(dir, full.names = FALSE, recursive = TRUE,
-                      pattern = "*.R$", ignore.case = TRUE)
+  files <- list.files(path = dir,
+                      full.names = TRUE,
+                      recursive = TRUE,
+                      pattern = suffix,
+                      ignore.case = TRUE)
 
-  is_found <- files %>%
-    vapply(readr::read_file, "character") %>%
-    grepl(pattern = pattern)
 
-  return(files[is_found])
+  is_found <- grepl(pattern = pattern,
+                    x = vapply(X = files,
+                               FUN = function(x) {
+                                 paste0(readLines(con = x, warn = FALSE),
+                                        collapse = "\n")
+                               },
+                               FUN.VALUE = "character"))
+
+
+  output <- gsub(x = files, pattern = paste0("^", dir), replacement = "")
+
+  return(output[is_found])
 
 }